pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool.black]
line-length = 120
target-version = ["py39"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "pass",
    "if __name__ == .__main__.:",
    "nocov",
    "if TYPE_CHECKING:",
]
fail_under = 80
show_missing = true

[tool.coverage.run]
branch = true
omit = [
    "tests/*"
]

[tool.poetry]
authors = ["Dawid Kraczkowski <dawid.kraczkowski@kaizenreporting.com>"]
description = "Abstraction Layer for Amazon DynamoDB"
license = "MIT"
name = "amano"
version = "0.1.0"

[tool.poetry.dependencies]
mypy-boto3-dynamodb = "^1.24.36"
boto = "^2.49.0"
python = "^3.8"
chili = "^1.7.0"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
flake8 = "^5.0.4"
isort = "^5.10.1"
mypy = "^0.971"
pylint = "^2.14.5"
pytest = "^7.1.2"
pytest-cov = "^3.0.0"
safety = "^2.1.1"
toml-sort = "^0.20.0"
